requirements.txt
Flask==2.3.3
requests==2.31.0
requests_oauthlib==1.3.1
psycopg2-binary==2.9.9
python-telegram-bot==20.7
python-dotenv==1.0.0
logging-formatter-anticrlf==1.2.1
aiohttp==3.9.1
